consolidate services
