+++ app.py
@@ -144,15 +144,13 @@ def run_search():
     from dotenv import load_dotenv
     from src.database.manager import setup_database_tables
     from src.database.video_operations import save_videos_to_database, save_video_features_to_database
-    from src.youtube.search import search_youtube_videos_by_query
-    from src.youtube.details import get_video_details_from_youtube
-    from src.youtube.utils import remove_duplicate_videos
+    from src.services.youtube_service import YouTubeService
     from src.ml.feature_extraction import extract_all_features_from_video
     from src.config.search_config import get_search_queries
     import random
-    
+
     load_dotenv()
-    
+
     api_key = os.getenv('YOUTUBE_API_KEY')
     if not api_key:
         print("Error: YOUTUBE_API_KEY not found in environment variables")
@@ -163,7 +161,9 @@ def run_search():
 
     print("🔍 Searching for more videos...")
 
+    youtube_service = YouTubeService(api_key)
     all_queries = get_search_queries()
+
     if len(all_queries) > 5:
         search_queries = all_queries[5:]
     else:
@@ -175,11 +175,10 @@ def run_search():
     all_videos = []
     for query in search_queries:
         print(f"  Searching: {query}")
-        video_ids = search_youtube_videos_by_query(api_key, query, 10)
-        videos = get_video_details_from_youtube(api_key, video_ids)
+        videos = youtube_service.search_and_get_details(query, 10)
         all_videos.extend(videos)
 
-    unique_videos = remove_duplicate_videos(all_videos)
+    unique_videos = YouTubeService.remove_duplicate_videos(all_videos)
 
     if unique_videos:
         save_videos_to_database(unique_videos, db_path)
