--- src/youtube/__init__.py
@@ -1,1 +1,0 @@
-# YouTube API operations package
